purchase_request: fix cancel purchase_request_line
Cancelling a purchase request:

With odoo core, when a purchase order line is cancelled, the
`move_dest_ids`'s procure methods are updated, as seen here
https://github.com/odoo/odoo/blob/14.0/addons/purchase_stock/models/purchase.py#L109-L110

With purchase_request, we should have the same behavior, if a
purchase_request_line related to stock_moves is cancelled, those moves
shouldn't be `Waiting for another operation anymore`.

Set a purchase request to done:

With the current implementation, we are able to set a purchase request
to done, even if no product or not all products has not been ordered from it.
Meaning that the related stock move will hang forever waiting for the
purchase request to be finished, which will never happen.